requirements.txt
Django >=4.2,<4.3
djangorestframework
TTS
# celery >= 5.2,<5.3
# redis